requirements.txt
telethon==1.36.0
sentence-transformers==3.2.0
torch>=2.0.0
faiss-cpu>=1.9.0
Pillow==11.0.0
imagehash==4.3.1
requests==2.32.3
python-dotenv==1.0.1
beautifulsoup4==4.12.3
transformers>=4.41.0
easyocr>=1.7.0


